check_url/measurement/requirements.txt
Requests==2.25.1
kafka-python==2.0.2